# api/swagger.yaml
openapi: 3.0.3
info:
  title: Budget Planner API
  description: A comprehensive budget planning and expense tracking API
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/personal/webapp
    email: support@budgetplanner.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082
    description: Development server
  - url: http://instance-agent.subnet05071228.vcn05071228.oraclevcn.com:8082
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login attempt response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: User logout
      description: Invalidate current session
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /auth/register:
    post:
      tags:
        - Auth
      summary: User registration
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/update-password:
    post:
      tags:
        - Auth
      summary: Update password
      description: Update the current user's password
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: password updated successfully
        '400':
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session_id
      description: Session-based authentication
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@localhost
        created_at:
          type: string
          format: date-time
          example: "2025-08-08T21:56:13Z"
        last_login:
          type: string
          format: date-time
          example: "2025-08-08T21:57:54Z"
          nullable: true
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login successful"
        session_id:
          type: string
          example: "cd69d274b34382c42862a929b7372b8a974327b81fe0c7010b8cc3f8513b3cd9"
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: newuser
        password:
          type: string
          format: password
          example: newpassword
        email:
          type: string
          format: email
          example: newuser@example.com
    UpdatePasswordRequest:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          format: password
          example: currentpassword
        new_password:
          type: string
          format: password
          example: newpassword
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"